+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/reconnection/SubgraphReconnector.java
@@ -32,7 +32,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
-package org.preesm.model.pisdf.util;
+package org.preesm.model.pisdf.reconnection;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -57,6 +57,7 @@ import org.preesm.model.pisdf.Fifo;
 import org.preesm.model.pisdf.Parameter;
 import org.preesm.model.pisdf.PiGraph;
 import org.preesm.model.pisdf.Port;
+import org.preesm.model.pisdf.util.PiMMSwitch;
 
 /**
  * Parse and connect hierarchical sub-{@link PiGraph} to a top level {@link PiGraph}.
@@ -98,15 +99,6 @@ public class SubgraphReconnector extends PiMMSwitch<Boolean> {
   /** The graph replacements. */
   private final Map<PiGraph, List<ActorByGraphReplacement>> graphReplacements = new LinkedHashMap<>();
 
-  /**
-   * Gets the graph replacements.
-   *
-   * @return the graph replacements
-   */
-  public Map<PiGraph, List<ActorByGraphReplacement>> getGraphReplacements() {
-    return this.graphReplacements;
-  }
-
   /** The current graph. */
   private PiGraph currentGraph = null;
 
@@ -116,7 +108,7 @@ public class SubgraphReconnector extends PiMMSwitch<Boolean> {
    * @param pg
    *          the graph process
    */
-  public void connectSubgraphs(final PiGraph pg) {
+  private void connectSubgraphs(final PiGraph pg) {
     doSwitch(pg);
     // Replace Actors with refinement by PiGraphs in pg and all its
     // subgraphs
@@ -270,7 +262,7 @@ public class SubgraphReconnector extends PiMMSwitch<Boolean> {
    * @param subGraph
    *          the subgraph linked to the hierarchical actor
    */
-  public static void reconnectPiGraph(final Actor hierarchicalActor, final PiGraph subGraph) {
+  private static void reconnectPiGraph(final Actor hierarchicalActor, final PiGraph subGraph) {
     SubgraphOriginalActorTracker.trackOriginalActor(hierarchicalActor, subGraph);
     SubgraphReconnector.reconnectDataInputPorts(hierarchicalActor, subGraph);
     SubgraphReconnector.reconnectDataOutputPorts(hierarchicalActor, subGraph);
@@ -369,5 +361,10 @@ public class SubgraphReconnector extends PiMMSwitch<Boolean> {
   private static void error(final Actor hierarchicalActor, final PiGraph subGraph, final Port port) {
     throw new PreesmRuntimeException("PiGraph '" + subGraph.getName() + "' does not have a corresponding "
         + port.getClass().getSimpleName() + " named '" + port.getName() + "' for Actor " + hierarchicalActor.getName());
+  }
+
+  public static void reconnectChildren(PiGraph graph) {
+    final SubgraphReconnector connector = new SubgraphReconnector();
+    connector.connectSubgraphs(graph);
   }
 }