+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/serialize/PiSDFExporterTask.java
@@ -52,6 +52,7 @@ import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.Path;
 import org.eclipse.emf.common.util.URI;
+import org.eclipse.emf.ecore.util.EcoreUtil;
 import org.preesm.commons.doc.annotations.Parameter;
 import org.preesm.commons.doc.annotations.Port;
 import org.preesm.commons.doc.annotations.PreesmTask;
@@ -59,6 +60,7 @@ import org.preesm.commons.doc.annotations.Value;
 import org.preesm.commons.exceptions.PreesmRuntimeException;
 import org.preesm.commons.files.WorkspaceUtils;
 import org.preesm.model.pisdf.PiGraph;
+import org.preesm.model.pisdf.reconnection.SubgraphDisconnector;
 import org.preesm.workflow.elements.Workflow;
 import org.preesm.workflow.implement.AbstractTaskImplementation;
 import org.preesm.workflow.implement.AbstractWorkflowNodeImplementation;
@@ -109,7 +111,13 @@ public class PiSDFExporterTask extends AbstractTaskImplementation {
     final String osString = documentFile.getLocation().toOSString();
     try (final OutputStream outStream = new FileOutputStream(osString);) {
       // Write the Graph to the OutputStream using the Pi format
-      new PiWriter(uri).write(graph, outStream);
+
+      final EcoreUtil.Copier copier = new EcoreUtil.Copier(false, false);
+      final PiGraph copy = (PiGraph) copier.copy(graph);
+      copier.copyReferences();
+
+      SubgraphDisconnector.disconnectSubGraphs(copy);
+      new PiWriter(uri).write(copy, outStream);
     } catch (IOException e) {
       throw new PreesmRuntimeException("Could not open outputstream file " + xmlPath.toString());
     }